BACKEND/requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.2
python-dotenv>=1.0
celery>=5.3
redis>=5.0
django-cors-headers>=4.4
gunicorn>=22.0
drf-spectacular>=0.27
argon2-cffi>=23.0
django-unfold>=0.10.0
